+++ game/squadrons/squadron.py
@@ -18,7 +18,7 @@ from game.theater import ParkingType
 from .pilot import Pilot, PilotStatus
 from ..db.database import Database
 from ..radio.radios import RadioFrequency
-from ..utils import meters
+from ..utils import meters, nautical_miles
 
 if TYPE_CHECKING:
     from game import Game
@@ -307,9 +307,21 @@ class Squadron:
 
         if ignore_range:
             return True
-        
+
         distance_to_target = meters(location.distance_to(self.location))
-        return distance_to_target <= self.aircraft.max_mission_range
+        max_plane_dist = nautical_miles(
+            self.coalition.game.settings.max_mission_range_planes
+        )
+        max_heli_dist = nautical_miles(
+            self.coalition.game.settings.max_mission_range_helicopters
+        )
+        if self.aircraft.helicopter:
+            return distance_to_target <= max(
+                self.aircraft.max_mission_range, max_heli_dist
+            )
+        return distance_to_target <= max(
+            self.aircraft.max_mission_range, max_plane_dist
+        )
 
     def operates_from(self, control_point: ControlPoint) -> bool:
         if not control_point.can_operate(self.aircraft):